General clean-up
Functionality has not been reviewed or changed.

- Fixed Parser call from Special:FormEdit/DemoAjax1 to "This breaks Ask
from SemanticMediawiki. If you use this feature, you need to test
whether you parser function work in ajax environment or not."
