+++ .gitignore
@@ -1,0 +1,11 @@
+*~
+*.kate-swp
+
+vendor/
+extensions/
+
+composer.phar
+composer.lock
+
+!.*
+.idea/
